+++ db/db_mysql/migrations/20260720000002_add_imap_capture_reply_body.sql
@@ -1,0 +1,5 @@
+-- +goose Up
+ALTER TABLE imap ADD COLUMN capture_reply_body BOOLEAN NOT NULL DEFAULT 1;
+
+-- +goose Down
+ALTER TABLE imap DROP COLUMN capture_reply_body;